Feature/cc 2223/dev decimal stock

+++ Bundles/ProductPackagingUnit/src/Spryker/Zed/ProductPackagingUnit/Dependency/Facade/ProductPackagingUnitToAvailabilityFacadeBridge.php
@@ -28,9 +28,9 @@ class ProductPackagingUnitToAvailabilityFacadeBridge implements ProductPackaging
      * @param string $sku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
-     * @return int
+     * @return float
      */
-    public function calculateStockForProductWithStore(string $sku, StoreTransfer $storeTransfer): int
+    public function calculateStockForProductWithStore(string $sku, StoreTransfer $storeTransfer): float
     {
         return $this->availabilityFacade->calculateStockForProductWithStore($sku, $storeTransfer);
     }
